fix the filtered count
